Move Processing widget context related classes out to their own file
