+++ python/PyQt6/gui/auto_additions/qgsprocessingwidgetcontext.py
@@ -1,0 +1,10 @@
+# The following has been generated automatically from src/gui/processing/qgsprocessingwidgetcontext.h
+try:
+    QgsProcessingWidgetContextGenerator.__abstract_methods__ = ['createWidgetContext']
+    QgsProcessingWidgetContextGenerator.__group__ = ['processing']
+except (NameError, AttributeError):
+    pass
+try:
+    QgsProcessingParameterWidgetContext.__group__ = ['processing']
+except (NameError, AttributeError):
+    pass
