+++ python/PyQt6/gui/auto_generated/processing/qgsprocessingwidgetcontext.sip.in
@@ -1,0 +1,242 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/processing/qgsprocessingwidgetcontext.h                      *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.py again   *
+ ************************************************************************/
+
+
+
+
+
+
+
+
+class QgsProcessingParameterWidgetContext
+{
+%Docstring(signature="appended")
+Contains settings which reflect the context in which a Processing
+parameter widget is shown.
+
+For instance, the parent model algorithm, a linked map canvas, and other
+relevant information which allows the widget to fine-tune its behavior.
+
+.. versionadded:: 3.4
+%End
+
+%TypeHeaderCode
+#include "qgsprocessingwidgetcontext.h"
+%End
+  public:
+    QgsProcessingParameterWidgetContext();
+
+    void setMapCanvas( QgsMapCanvas *canvas );
+%Docstring
+Sets the map ``canvas`` associated with the widget. This allows the
+widget to retrieve the current map scale and other properties from the
+canvas.
+
+.. seealso:: :py:func:`mapCanvas`
+%End
+
+    QgsMapCanvas *mapCanvas() const;
+%Docstring
+Returns the map canvas associated with the widget.
+
+.. seealso:: :py:func:`setMapCanvas`
+%End
+
+    void setMessageBar( QgsMessageBar *bar );
+%Docstring
+Sets the message ``bar`` associated with the widget. This allows the
+widget to push feedback messages to the user.
+
+.. seealso:: :py:func:`messageBar`
+
+.. versionadded:: 3.12
+%End
+
+    QgsMessageBar *messageBar() const;
+%Docstring
+Returns the message bar associated with the widget. This allows the
+widget to push feedback messages to the user.
+
+.. seealso:: :py:func:`setMessageBar`
+
+.. versionadded:: 3.12
+%End
+
+    void setBrowserModel( QgsBrowserGuiModel *model );
+%Docstring
+Sets the browser ``model`` associated with the widget. This will usually
+be the shared app instance of the browser model
+
+.. seealso:: :py:func:`browserModel`
+
+.. versionadded:: 3.14
+%End
+
+    QgsBrowserGuiModel *browserModel() const;
+%Docstring
+Returns the browser model associated with the widget.
+
+.. seealso:: :py:func:`setBrowserModel`
+
+.. versionadded:: 3.12
+%End
+
+    void setProject( QgsProject *project );
+%Docstring
+Sets the ``project`` associated with the widget. This allows the widget
+to retrieve the map layers and other properties from the correct
+project.
+
+.. seealso:: :py:func:`project`
+
+.. versionadded:: 3.8
+%End
+
+    QgsProject *project() const;
+%Docstring
+Returns the project associated with the widget.
+
+.. seealso:: :py:func:`setProject`
+%End
+
+    QgsProcessingModelAlgorithm *model() const;
+%Docstring
+Returns the model which the parameter widget is associated with.
+
+.. seealso:: :py:func:`setModel`
+
+.. seealso:: :py:func:`modelChildAlgorithmId`
+%End
+
+    void setModel( QgsProcessingModelAlgorithm *model );
+%Docstring
+Sets the ``model`` which the parameter widget is associated with.
+
+.. seealso:: :py:func:`model`
+
+.. seealso:: :py:func:`setModelChildAlgorithmId`
+%End
+
+    QString modelChildAlgorithmId() const;
+%Docstring
+Returns the child algorithm ID within the model which the parameter
+widget is associated with.
+
+.. seealso:: :py:func:`setModelChildAlgorithmId`
+
+.. seealso:: :py:func:`model`
+%End
+
+    void setModelChildAlgorithmId( const QString &id );
+%Docstring
+Sets the child algorithm ``id`` within the model which the parameter
+widget is associated with.
+
+.. seealso:: :py:func:`modelChildAlgorithmId`
+
+.. seealso:: :py:func:`setModel`
+%End
+
+    QgsMapLayer *activeLayer() const;
+%Docstring
+Returns the current active layer.
+
+.. seealso:: :py:func:`setActiveLayer`
+
+.. versionadded:: 3.14
+%End
+
+    void setActiveLayer( QgsMapLayer *layer );
+%Docstring
+Sets the current active ``layer``.
+
+.. seealso:: :py:func:`activeLayer`
+
+.. versionadded:: 3.14
+%End
+
+    void registerProcessingContextGenerator( QgsProcessingContextGenerator *generator );
+%Docstring
+Registers a Processing context ``generator`` class that will be used to
+retrieve a Processing context for the widget when required.
+
+The ``generator`` must exist for the lifetime of the widget, ownership
+is not transferred.
+
+.. seealso:: :py:func:`processingContextGenerator`
+
+.. versionadded:: 4.0
+%End
+
+    QgsProcessingContextGenerator *processingContextGenerator();
+%Docstring
+Returns the Processing context generator class that will be used to
+retrieve a Processing context for the widget when required.
+
+.. seealso:: :py:func:`registerProcessingContextGenerator`
+
+.. versionadded:: 4.0
+%End
+
+    QgsModelDesignerDialog *modelDesignerDialog() const;
+%Docstring
+Returns the associated model designer dialog, if applicable.
+
+.. warning::
+
+   This method is not considered stable API
+
+.. seealso:: :py:func:`setModelDesignerDialog`
+
+.. versionadded:: 4.0
+%End
+
+    void setModelDesignerDialog( QgsModelDesignerDialog *dialog );
+%Docstring
+Sets the associated model designer ``dialog``, if applicable.
+
+.. warning::
+
+   This method is not considered stable API
+
+.. seealso:: :py:func:`modelDesignerDialog`
+
+.. versionadded:: 4.0
+%End
+
+};
+
+class QgsProcessingWidgetContextGenerator
+{
+%Docstring(signature="appended")
+An interface for objects which can create Processing widget contexts.
+
+.. versionadded:: 4.4
+%End
+
+%TypeHeaderCode
+#include "qgsprocessingwidgetcontext.h"
+%End
+  public:
+    virtual QgsProcessingParameterWidgetContext createWidgetContext() = 0;
+%Docstring
+This method needs to be reimplemented in all classes which implement
+this interface and return a Processing widget context.
+%End
+
+    virtual ~QgsProcessingWidgetContextGenerator();
+};
+
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * src/gui/processing/qgsprocessingwidgetcontext.h                      *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.py again   *
+ ************************************************************************/
